Point 제거, primitive type 사용
