+++ src/main/java/ladder/domain/Line.java
@@ -3,48 +3,61 @@ package ladder.domain;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Random;
 import java.util.stream.IntStream;
 
 public class Line {
     private static final int STARTING_POINT = 0;
-    private List<Point> points;
+    private List<Boolean> points;
 
-    public Line(List<Point> points) {
+    public Line(List<Boolean> points) {
         this.points = points;
     }
 
     public static Line generate(int height) {
-        List<Point> generatedPoints = new ArrayList<>();
+        List<Boolean> generatedPoints = new ArrayList<>();
         IntStream.range(0, height)
             .forEach(i -> generatedPoints.add(fillCurrent(generatedPoints, i)));
         return new Line(generatedPoints);
     }
 
-    private static Point fillCurrent(List<Point> generatedPoints, int position) {
+    private static Boolean fillCurrent(List<Boolean> generatedPoints, int position) {
         if (position == STARTING_POINT)
-            return new Point(false);
+            return false;
 
-        Point randomPoint = Point.generateRandom();
+        Boolean randomBoolean = new Random().nextBoolean();
 
-        if (isOverlap(generatedPoints.get(position - 1), randomPoint)) {
-            return Point.reverse(randomPoint);
+        if (isOverlap(generatedPoints.get(position - 1), randomBoolean)) {
+            return !randomBoolean;
         }
 
-        return randomPoint;
+        return randomBoolean;
     }
 
-    static boolean isOverlap(Point beforeBoolean, Point randomBoolean) {
-        return Point.isTrueOverlap(beforeBoolean, randomBoolean);
+    static boolean isOverlap(Boolean beforeBoolean, Boolean randomBoolean) {
+        return beforeBoolean && randomBoolean;
     }
 
     @Override
     public String toString() {
         StringBuilder builder = new StringBuilder();
-        this.points.forEach(point -> builder.append(point).append("l"));
+        this.points.forEach(point -> builder.append(beautify(point)).append("l"));
         return builder.toString();
     }
 
-    public List<Point> getPoints() {
+    public static String beautify(Boolean point) {
+        StringBuilder builder = new StringBuilder();
+        String fill = " ";
+        if (point) {
+            fill = "-";
+        }
+        for (int i = 0; i < Person.MAX_NAME_LENGTH; i++) {
+            builder.append(fill);
+        }
+        return builder.toString();
+    }
+
+    public List<Boolean> getPoints() {
         return Collections.unmodifiableList(points);
     }
 }
