+++ src/test/java/ladder/domain/LineTest.java
@@ -16,13 +16,23 @@ public class LineTest {
 
     @Test
     public void 라인이_겹칠때() {
-        assertTrue(Line.isOverlap(new Point(true), new Point(true)));
+        assertTrue(Line.isOverlap(true, true));
     }
 
     @Test
     public void 라인이_안겹칠때() {
-        assertFalse(Line.isOverlap(new Point(true), new Point(false)));
-        assertFalse(Line.isOverlap(new Point(false), new Point(true)));
-        assertFalse(Line.isOverlap(new Point(false), new Point(false)));
+        assertFalse(Line.isOverlap(true, false));
+        assertFalse(Line.isOverlap(false, true));
+        assertFalse(Line.isOverlap(false, false));
+    }
+
+    @Test
+    public void 라인을_그리지_않는다() {
+        assertThat(Line.beautify(false)).isEqualTo("     ");
+    }
+
+    @Test
+    public void 라인을_그린다() {
+        assertThat(Line.beautify(true)).isEqualTo("-----");
     }
 }
